refactor(node-runtime): share the worker thread transport
Voice transcription needs the same worker plumbing the semantic index already
has (tsx bootstrap, correlated postMessage, graceful runtime close before
terminate), plus a transfer list so PCM buffers are handed over rather than
copied. Move the feature-agnostic half to worker/thread-transport.ts and keep
createSemanticIndexWorkerThreadTransport as the semantic index's entry point;
its behaviour, options and exported types are unchanged.
